requirement.txt
easydict
numpy
torch
tensorboard
nltk
prettytable
PyYAML
matplotlib
scikit-learn